requirements.txt
asgiref==3.11.0
boto3==1.42.2
botocore==1.42.2
dj-database-url==3.0.1
Django==5.1.3
django-cors-headers==4.3.1
django-debug-toolbar==6.1.0
djangorestframework==3.15.2
environs==14.5.0
jmespath==1.0.1
marshmallow==4.1.0
pillow==10.4.0
psycopg==3.2.13
psycopg-binary==3.2.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.16.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
